fixed temp min and temp max, added utils toPercent, removed id column from CityEntity

+++ main-screen/src/main/java/com/weather/main/screen/model/WeatherMetricsPres.kt
@@ -10,5 +10,5 @@ data class WeatherMetricsPres(
     /** Облачность, в %*/
     val cloudiness: Int,
     /** Вероятность осадков, в %*/
-    val pop: Double,
+    val pop: Int,
 )
